+++ src/Mvx.ApiClient.Net/EndpointPaths.cs
@@ -1,17 +1,17 @@
-﻿namespace Mvx.ApiClient.Net;
+namespace Mvx.ApiClient.Net;
 
 internal static class EndpointPaths
 {
-    internal const string MexEconomics = "/mex/economics";
-    internal const string MexPairs = "/mex/pairs";
-    internal const string MexPairsCount = "/mex/pairs/count";
-    internal const string MexPairDetails = "/mex/pairs/{0}/{1}";
-    internal const string MexTokens = "/mex/tokens";
-    internal const string MexTokensCount = "/mex/tokens/count";
-    internal const string MexTokenDetails = "/mex/tokens/{0}";
-    internal const string MexFarms = "/mex/farms";
-    internal const string MexFarmsCount = "/mex/farms/count";
-    
+    internal const string XExchangeEconomics = "/mex/economics";
+    internal const string XExchangePairs = "/mex/pairs";
+    internal const string XExchangePairsCount = "/mex/pairs/count";
+    internal const string XExchangePairDetails = "/mex/pairs/{0}/{1}";
+    internal const string XExchangeTokens = "/mex/tokens";
+    internal const string XExchangeTokensCount = "/mex/tokens/count";
+    internal const string XExchangeTokenDetails = "/mex/tokens/{0}";
+    internal const string XExchangeFarms = "/mex/farms";
+    internal const string XExchangeFarmsCount = "/mex/farms/count";
+
     internal const string NetworkStats = "/stats";
     internal const string NetworkEconomics = "/economics";
     internal const string NetworkConstants = "/constants";