+++ src/Mvx.ApiClient.Net/Models/XExchange/XExchangeEconomicsDto.cs
@@ -1,32 +1,26 @@
 using System.Text.Json.Serialization;
 
-namespace Mvx.ApiClient.Net.Models.Mex;
+namespace Mvx.ApiClient.Net.Models.XExchange;
 
 /// <summary>
-/// xExchange economics
+/// xExchange economics information.
 /// </summary>
-/// <param name="TotalSupply">Total supply of MEX</param>
-/// <param name="CirculatingSupply">Circulating supply of MEX</param>
-/// <param name="Price">Current MEX price</param>
-/// <param name="MarketCap">Current market cap of MEX</param>
-/// <param name="Volume">Volume (24h) of MEX</param>
-/// <param name="MarketPairs">Number of market pairs</param>
-public record MexEconomicsDto(
+public sealed record XExchangeEconomicsDto(
     [property: JsonPropertyName("totalSupply")]
     long TotalSupply,
-    
+
     [property: JsonPropertyName("circulatingSupply")]
     long CirculatingSupply,
-    
+
     [property: JsonPropertyName("price")]
     double Price,
-    
+
     [property: JsonPropertyName("marketCap")]
     long MarketCap,
-    
+
     [property: JsonPropertyName("volume24h")]
     long Volume,
-    
+
     [property: JsonPropertyName("marketPairs")]
     int MarketPairs
 );