+++ src/Mvx.ApiClient.Net/Models/XExchange/XExchangePairDto.cs
@@ -1,108 +1,83 @@
 using System.Text.Json.Serialization;
 
-namespace Mvx.ApiClient.Net.Models.Mex;
+namespace Mvx.ApiClient.Net.Models.XExchange;
 
 /// <summary>
-/// xExchange Pair
+/// xExchange pair.
 /// </summary>
-/// <param name="Address"></param>
-/// <param name="Id"></param>
-/// <param name="Symbol"></param>
-/// <param name="Name"></param>
-/// <param name="Price"></param>
-/// <param name="BasePreviousPrice"></param>
-/// <param name="QuotePreviousPrice"></param>
-/// <param name="BaseId"></param>
-/// <param name="BaseSymbol"></param>
-/// <param name="BaseName"></param>
-/// <param name="BasePrice"></param>
-/// <param name="QuoteId"></param>
-/// <param name="QuoteSymbol"></param>
-/// <param name="QuoteName"></param>
-/// <param name="QuotePrice"></param>
-/// <param name="TotalValue"></param>
-/// <param name="Volume"></param>
-/// <param name="State"></param>
-/// <param name="Type"></param>
-/// <param name="Exchange"></param>
-/// <param name="HasFarms"></param>
-/// <param name="HasDualFarms"></param>
-/// <param name="TradesCount"></param>
-/// <param name="PreviousTradesCount"></param>
-/// <param name="DeployedAt"></param>
-public record MexPairDto(
+public sealed record XExchangePairDto(
     [property: JsonPropertyName("address")]
     string Address,
-    
+
     [property: JsonPropertyName("id")]
     string Id,
-    
+
     [property: JsonPropertyName("symbol")]
     string Symbol,
-    
+
     [property: JsonPropertyName("name")]
     string Name,
-    
+
     [property: JsonPropertyName("price")]
     double Price,
-    
+
     [property: JsonPropertyName("basePrevious24hPrice")]
     double BasePreviousPrice,
-    
+
     [property: JsonPropertyName("quotePrevious24hPrice")]
     double QuotePreviousPrice,
-    
+
     [property: JsonPropertyName("baseId")]
     string BaseId,
-    
+
     [property: JsonPropertyName("baseSymbol")]
     string BaseSymbol,
-    
+
     [property: JsonPropertyName("baseName")]
     string BaseName,
-    
+
     [property: JsonPropertyName("basePrice")]
     double BasePrice,
-    
+
     [property: JsonPropertyName("quoteId")]
     string QuoteId,
-    
+
     [property: JsonPropertyName("quoteSymbol")]
     string QuoteSymbol,
-    
+
     [property: JsonPropertyName("quoteName")]
     string QuoteName,
-    
+
     [property: JsonPropertyName("quotePrice")]
     double QuotePrice,
-    
+
     [property: JsonPropertyName("totalValue")]
     double TotalValue,
-    
+
     [property: JsonPropertyName("volume24h")]
     double Volume,
-    
+
     [property: JsonPropertyName("state")]
-    MexPairState State,
-    
+    XExchangePairState State,
+
     [property: JsonPropertyName("type")]
-    MexPairType Type,
-    
+    XExchangePairType Type,
+
     [property: JsonPropertyName("exchange")]
     string Exchange,
-    
+
     [property: JsonPropertyName("hasFarms")]
     bool HasFarms,
-    
+
     [property: JsonPropertyName("hasDualFarms")]
     bool HasDualFarms,
-    
+
     [property: JsonPropertyName("tradesCount")]
     long TradesCount,
-    
+
     [property: JsonPropertyName("tradesCount24h")]
     long PreviousTradesCount,
-    
+
     [property: JsonPropertyName("deployedAt")]
     long DeployedAt
 );
@@ -110,7 +85,7 @@ public record MexPairDto(
 /// <summary>
 /// xExchange pair state.
 /// </summary>
-public enum MexPairState
+public enum XExchangePairState
 {
     /// <summary>
     /// Active pair.
@@ -136,7 +111,7 @@ public enum MexPairState
 /// <summary>
 /// xExchange pair category.
 /// </summary>
-public enum MexPairType
+public enum XExchangePairType
 {
     /// <summary>
     /// Core xExchange pair.